requirements.txt
fastapi
uvicorn
SQLAlchemy
pydantic
alembic
requests
beautifulsoup4
python-dotenv
sqlalchemy[asyncio]
asyncpg
aiosqlite
